// packages/weapp-tailwindcss/src/bundlers/vite/generate-bundle/types.ts
export interface GenerateBundleThis {
  addWatchFile?: (id: string) => void
  emitFile?: (emittedFile: {
    type: 'asset'
    fileName: string
    source: string
  }) => string
  getModuleInfo?: (id: string) => { code: string | null } | null
}